refactoring for seperate loader files.
